Allow overwriting the generated base name
Purpose: Prior to this change the airtable client code was always generated in a path containing the actual Airtable base name. This was preventing users from applying structure changes to a duplicated base and generating code from that duplicate. This change allows to use the name configuration property to overwrite the generated code's base name.

Changes:

Map the config base name instead of the actual name in case it is provided
Readme update
Bump up package version